Agents/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
supabase==2.0.3
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
transformers==4.35.2
scikit-learn==1.3.2
spacy==3.7.2
pytesseract==0.3.10
tldextract==5.1.1
prometheus-client==0.19.0
pydantic==2.5.2
pytest==7.4.3
python-multipart==0.0.6
aiohttp==3.9.1 